ml/nifty500/requirements.txt
smartapi-python
pyotp
pandas
pyarrow
requests
logzero
websocket-client
NorenRestApiPy
feedparser
google-generativeai
python-telegram-bot
python-dotenv
yfinance
google-cloud-storage
xgboost
scikit-learn
numpy
